Fixes #91.  Make IO resources not autoclose.
A later JRuby release may allow these autoclose=false statements to be
removed (although @handle may desire to keep it's due to to_io handing
that IO out to external consumers).  The basic problem is currently during
shutdown JRuby Threads may not have finished executing before finalizers
start firing.  JRuby is looking at sending a more explicit termination
request similar to MRI to threads on shutdown, which may obviate the need
for this.  That change is too risky in the near term so we are requesting
this PR for the time being.